Rewrite of most of RimworldDiscoverTechsClass.cs. New ModExtension for ResearchProjectDef, and edits to ResearchProjects_Blueprints.xml to use the extension.
No longer creates dynamic ResearchProjectDef prerequisites.

Instead, uses Harmony to patch ResearchProjectDef.PrerequisitesComplete to check for the ModExtension, then performs a custom prereq check. Modifies result accordingly. If the property is being accessed for a research project that isn't one of ours, it runs the original code for best compatibility.